requirements.txt
# Servidor ASR Qwen3
# Para Mac ARM64 instalar torch con:
#   pip install torch torchaudio --index-url https://download.pytorch.org/whl/cpu
# Para Linux x86_64 CPU:
#   pip install torch torchaudio --index-url https://download.pytorch.org/whl/cpu

# Core framework
torch==2.11.0
torchaudio==2.11.0
qwen-asr==0.0.6

# API
fastapi==0.136.1
uvicorn[standard]==0.46.0
python-multipart==0.0.27

# Diarización
pyannote.audio==4.0.4

# Config
python-dotenv==1.2.2

# Transitive critical versions (pinned for stability)
transformers==4.57.6
accelerate==1.12.0
numpy==2.4.3
librosa==0.11.0
soundfile==0.13.1
